"Some general changes done, comments added."

+++ src/main/java/com/green/samplecompany2/repositories/ComputerRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ComputerRepository extends JpaRepository<Computer, Long> {
+public interface ComputerRepository extends JpaRepository<Computer, Integer> {
 
 
 }
